+++ src/OpenLoco/src/Objects/CompetitorObject.cpp
@@ -11,10 +11,6 @@
 
 namespace OpenLoco
 {
-    static loco_global<uint8_t, 0x0112C211> _intelligence;    // return of loadTemporaryObject
-    static loco_global<uint8_t, 0x0112C212> _aggressiveness;  // return of loadTemporaryObject
-    static loco_global<uint8_t, 0x0112C213> _competitiveness; // return of loadTemporaryObject
-
     // TODO: Should only be defined in ObjectSelectionWindow
     static constexpr uint8_t kDescriptionRowHeight = 10;
     static constexpr Ui::Size kObjectPreviewSize = { 112, 112 };
@@ -107,12 +103,6 @@ namespace OpenLoco
 
         // Ensure we've loaded the entire object
         assert(remainingData.size() == imageRes.tableLength);
-
-        // Copy competitor stats to global
-        // TODO: Refactor to not pass by global!
-        _intelligence = intelligence;
-        _aggressiveness = aggressiveness;
-        _competitiveness = competitiveness;
     }
 
     // 0x00434D08