+++ src/OpenLoco/src/Ui/Windows/ScenarioSelect.cpp
@@ -136,7 +136,7 @@ namespace OpenLoco::Ui::Windows::ScenarioSelect
         initTabs(self);
 
         // Select the last tab used, or the first available one.
-        uint8_t selectedTab = Config::get().old.scenarioSelectedTab;
+        uint8_t selectedTab = Config::get().scenarioSelectedTab;
         if (self->widgets[widx::tab0 + selectedTab].hidden)
         {
             selectedTab = 0;
@@ -474,7 +474,7 @@ namespace OpenLoco::Ui::Windows::ScenarioSelect
 
                 self.currentTab = selectedCategory;
 
-                auto& config = Config::get().old;
+                auto& config = Config::get();
                 config.scenarioSelectedTab = selectedCategory;
                 Config::write();