+++ src/OpenLoco/src/Ui/Windows/StationWindow.cpp
@@ -252,7 +252,7 @@ namespace OpenLoco::Ui::Windows::Station
             }
             else
             {
-                if (Config::get().hasFlags(Config::Flags::gridlinesOnLandscape))
+                if (Config::get().gridlinesOnLandscape)
                 {
                     flags |= ViewportFlags::gridlines_on_landscape;
                 }