+++ src/OpenLoco/src/ViewportManager.cpp
@@ -85,7 +85,7 @@ namespace OpenLoco::Ui::ViewportManager
         vp->zoom = static_cast<uint8_t>(zoom);
         vp->flags = ViewportFlags::none;
 
-        if (OpenLoco::Config::get().hasFlags(Config::Flags::gridlinesOnLandscape))
+        if (Config::get().gridlinesOnLandscape)
         {
             vp->flags |= ViewportFlags::gridlines_on_landscape;
         }
@@ -301,7 +301,7 @@ namespace OpenLoco::Ui::ViewportManager
         rect.right = t->spriteRight;
         rect.bottom = t->spriteBottom;
 
-        auto level = (ZoomLevel)std::min(Config::get().old.vehiclesMinScale, (uint8_t)zoom);
+        auto level = (ZoomLevel)std::min(Config::get().vehiclesMinScale, (uint8_t)zoom);
         invalidate(rect, level);
     }